+++ modules/@apostrophecms/asset/lib/webpack/utils.js
@@ -62,19 +62,16 @@ module.exports = {
     };
   },
 
-  fillExtraBundles (verifiedBundles = []) {
-    const getFileName = (p) => p.substr(p.lastIndexOf('/') + 1)
-      .replace(/\.(js|scss)$/, '');
-
-    return verifiedBundles.reduce((acc, { paths }) => {
+  fillExtraBundles (verifiedBundles = {}) {
+    return Object.entries(verifiedBundles).reduce((acc, [ bundleName, { js, scss } ]) => {
       return {
         js: [
           ...acc.js,
-          ...paths.filter((p) => p.endsWith('.js')).map((p) => getFileName(p))
+          ...(js.length && !acc.js.includes(bundleName)) ? [ bundleName ] : []
         ],
         css: [
           ...acc.css,
-          ...paths.filter((p) => p.endsWith('.scss')).map((p) => getFileName(p))
+          ...(scss.length && !acc.css.includes(bundleName)) ? [ bundleName ] : []
         ]
       };
     }, {
@@ -99,6 +96,54 @@ module.exports = {
         ...es5Bundles
       ];
     }, []);
+  },
+
+  writeBundlesImportFiles ({
+    name,
+    buildDir,
+    mainBundleName,
+    verifiedBundles,
+    getImportFileOutput,
+    writeImportFile
+  }) {
+    if (!name.includes('src')) {
+      return [];
+    }
+
+    const bundlesOutputs = Object.entries(verifiedBundles)
+      .map(([ bundleName, paths ]) => {
+        return {
+          bundleName,
+          importFile: `${buildDir}/${bundleName}-import.js`,
+          js: getImportFileOutput(paths.js, {
+            invokeApps: true,
+            enumerateImports: true,
+            requireDefaultExport: true
+          }),
+          scss: getImportFileOutput(paths.scss, {
+            enumerateImports: true,
+            importSuffix: 'Stylesheet'
+          })
+        };
+      });
+
+    for (const output of bundlesOutputs) {
+      writeImportFile({
+        importFile: output.importFile,
+        indexJs: output.js,
+        indexSass: output.scss
+      });
+    }
+
+    return bundlesOutputs.reduce((acc, { bundleName, importFile }) => {
+      return {
+        ...acc,
+        [bundleName]: {
+          import: importFile,
+          dependOn: mainBundleName
+        }
+      };
+    }, {});
   }
 };
 
@@ -153,17 +198,36 @@ async function verifyBundlesEntryPoints (bundles) {
 
     return {
       bundleName,
-      paths: [
-        ...jsFileExists ? [ jsPath ] : [],
-        ...scssFileExists ? [ scssPath ] : []
-      ]
+      ...jsFileExists && { jsPath: jsPath },
+      ...scssFileExists && { scssPath: scssPath }
     };
   });
 
-  const bundlesPaths = (await Promise.all(checkPathsPromises))
-    .filter((bundle) => bundle.paths.length);
+  const bundlesPaths = await Promise.all(checkPathsPromises);
 
-  return bundlesPaths;
+  const packedFilesByBundle = bundlesPaths.reduce((acc, {
+    bundleName, jsPath, scssPath
+  }) => {
+    if (!jsPath && !scssPath) {
+      return acc;
+    }
+
+    return {
+      ...acc,
+      [bundleName]: {
+        js: [
+          ...acc[bundleName] ? acc[bundleName].js : [],
+          ...jsPath ? [ jsPath ] : []
+        ],
+        scss: [
+          ...acc[bundleName] ? acc[bundleName].scss : [],
+          ...scssPath ? [ scssPath ] : []
+        ]
+      }
+    };
+  }, {});
+
+  return packedFilesByBundle;
 };
 
 function formatConfigs (chain, webpackConfigs) {